requirements.txt
opencv_python>=4.2.0.34
matplotlib>=3.1.2
numpy>=1.17.3
pywin32>=228
PyYAML>=5.3.1
